+++ Echo/src/utopia/echo/model/response/chat/ReplyMessage.scala
@@ -1,14 +1,14 @@
 package utopia.echo.model.response.chat
 
 import utopia.echo.model.enumeration.ChatRole
-import utopia.echo.model.response.OllamaResponse
+import utopia.echo.model.response.{OllamaResponse, OllamaResponseLike}
 
 /**
   * Common trait for chat responses, whether buffered or streamed
   * @author Mikko Hilpinen
   * @since 21.07.2024, v1.0
   */
-trait ReplyMessage extends OllamaResponse[BufferedReplyMessage]
+trait ReplyMessage extends OllamaResponse with OllamaResponseLike[BufferedReplyMessage]
 {
 	/**
 	  * @return Role of the sender of this message